Created blog index page
Some of the formatting from the tutorial I don't quite understand. What is the » symbol used for? Why is everything following the {% for post in site.posts %} indented? Is that a comment?

Reformatted it a little myself to something that seems to make a bit more sense... let's see if it works properly.